+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix `PARTITION` syntax errors (#377)
 - Fix `ALTER USER` when used with `IDENTIFIED WITH/VIA/BY` option (#431)
 - Fix `COALESCE PARTITION` in `ALTER TABLE`, rather than `COALESCE` (#323)
+- Support `ALGORITHM` and `LOCK` options in `ALTER TABLE` statements (#319)
 
 ## [5.7.0] - 2023-01-25